requirements.txt
python-dotenv
flask
flask-restx
flask_socketio
flask-sqlalchemy
flask-jwt-extended
mysql-connector-python
schedule
apscheduler
Flask-Mail
pyotp
flask-cors